refactor/FE: Extract Desktop product images component
